Break up login.svelte into smaller components and extract utils
